+++ test/e2e/test_cases/static_chart_types/cartesian_coo_sys/marimekko_rectangle_1dis_2con.mjs
@@ -15,7 +15,7 @@ const testSteps = [
 					y: 'Value 1 (+)',
 					x: ['Country', 'Value 2 (+)'],
 					color: 'Country',
-					label: ['Value 2 (+)', 'Country']
+					label: ['Value 2 (+)']
 				},
 				title: 'Mekko chart'
 			}